% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REMO.v1.GRCh38.CL_ID}
\alias{REMO.v1.GRCh38.CL_ID}
\title{GRCh38 REMO v1 cell ontology IDs}
\format{
A named list. The name of each element corresponds to a CL ID, and
the list element is a character vector of REMO module IDs.
}
\source{
\url{https://obophenotype.github.io/cell-ontology/}
}
\usage{
REMO.v1.GRCh38.CL_ID
}
\description{
Cell Ontology (CL) IDs associated with each REMO v1 GRCh38 module.
}
\author{
Tim Stuart

Chrysania Lim
}
\keyword{datasets}
