% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REMO.v1.GRCh38.metadata}
\alias{REMO.v1.GRCh38.metadata}
\title{GRCh38 REMO v1 metadata}
\format{
A data frame with the following columns:

\describe{
  \item{REMO}{REMO ID}
  \item{CREs}{Number of cis-regulatory elements in the module}
  \item{Bases}{Total number of genome bases covered by CREs in the module}
  \item{Chromosome}{Name of the chromosome with the module}
  \item{GC_mean}{Mean GC content of the module}
  \item{CL}{Cell Ontology terms associated with the module}
}
}
\source{
\url{https://doi.org/10.64898/2025.12.10.692786}

\doi{10.64898/2025.12.10.692786}
}
\usage{
REMO.v1.GRCh38.metadata
}
\description{
Metadata associated with each REMO v1 GRCh38 module.
}
\author{
Tim Stuart

Chrysania Lim
}
\keyword{datasets}
