% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REMO.v1.GRCh38}
\alias{REMO.v1.GRCh38}
\title{GRCh38 REMO v1 modules}
\format{
A GRanges object with the following metadata columns
\describe{
  \item{REMO}{REMO module membership}
}
}
\source{
\url{}

\doi{}
}
\usage{
REMO.v1.GRCh38
}
\description{
REMO module coordinates and module membership for the GRCh38 genome
}
\author{
Tim Stuart

Chrysania Lim
}
\keyword{datasets}
