% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REMO.v1.GRCh38.terms}
\alias{REMO.v1.GRCh38.terms}
\title{GRCh38 REMO v1 cell ontology terms}
\format{
A named list. The name of each element corresponds to a CL name, and
the list element is a character vector of REMO module IDs.
}
\source{
\url{}

\doi{}
}
\usage{
REMO.v1.GRCh38.terms
}
\description{
Cell Ontology (CL) terms associated with each REMO v1 GRCh38 module.
}
\author{
Tim Stuart

Chrysania Lim
}
\keyword{datasets}
