% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{REMO.v1.GRCh38.tissues}
\alias{REMO.v1.GRCh38.tissues}
\title{GRCh38 tissue cell types}
\format{
A named list. Each element is a character vector of CL names present
in the tissue.
}
\source{
\url{https://humanatlas.io/cell-type-annotations}

\doi{doi: 10.1038/s41588-024-01993-3}

\doi{doi: 10.1016/j.cell.2021.04.048}

\doi{doi: 10.1016/j.cell.2023.11.026}
}
\usage{
REMO.v1.GRCh38.tissues
}
\description{
Cell type terms present in each human tissue.
}
\details{
These lists were generated by combining information from the Human Reference
Atlas cell type annotations:

\describe{
 \item{Azimuth v1.2}{\url{https://doi.org/10.48539/HBM543.NVTF.775}}
 \item{CellTypist v1.1}{\url{https://doi.org/10.48539/HBM877.ZTNJ.972}}
 \item{popV v1.2}{\url{https://doi.org/10.48539/HBM866.LDBJ.546}}
}
}
\keyword{datasets}
